% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.tutorials-package.R
\docType{package}
\name{tidymodels.tutorials-package}
\alias{tidymodels.tutorials}
\alias{tidymodels.tutorials-package}
\title{tidymodels.tutorials: Tutorials for Tidy Modeling with R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Kane \email{dave.kane@gmail.com} (\href{https://orcid.org/0000-0002-6660-3934}{ORCID}) [copyright holder]

}
\keyword{internal}
